finished nesting in backend

+++ server/fileController.js
@@ -1,6 +1,9 @@
 const fs = require('fs-extra');
 const path = require('path');
+const { emitWarning } = require('process');
 const execSync = require('child_process').execSync;
+const ncp=require('ncp').ncp
+const glob = require('glob');
 
 
 const fileController = {
@@ -51,6 +54,9 @@ const fileController = {
   // },
 
   postFolder: function (req, res, next) {
+
+    const folderDir = 'server/ExportFolder/nextsketch/src/'
+
     if (req.body.name) {
       const dir = 'server/ExportFolder/nextsketch/src/app/';
       fs.mkdirSync(path.join(dir, req.body.name));
@@ -58,23 +64,41 @@ const fileController = {
     }
 
     if (req.body.fileName) {
-      const fileDir = "server/ExportFolder/nextsketch/src/app/" + req.body.folderName;
-      if(req.body.isFolder){
-        console.log('in middleware', req.body.isFolder)
+    
+    function recall(folderDir) {
+      // Get an array of all files and directories in the passed directory using fs.readdirSync
+      const fileList = fs.readdirSync(folderDir);
 
-        fs.mkdirSync(path.join(fileDir, req.body.fileName))
+      // Create the full path of the file/directory by concatenating the passed directory and file/directory name
+      for (const file of fileList) {
+        const name = `${folderDir}/${file}`;
+
+        //skip reading over all the files inside node_modules for efficiency
+        if (file === 'node_modules') {
+          continue;
+        }
+
+        if (file === req.body.folderName) { 
+          if (req.body.isFolder) fs.mkdirSync(path.join(name, req.body.fileName))
+          else fs.writeFileSync(path.join(name, req.body.fileName), '')
+          return;
+        }
+
+        // Check if the current file/directory is a directory using fs.statSync
+        if (fs.statSync(name).isDirectory()) {
+          // If it is a directory, recursively call the getFiles function with the directory path and the files array
+
+          recall(name);
+        }
       }
-      else fs.writeFileSync(path.join(fileDir, req.body.fileName), "");
-      return next();
+      return;
     }
+    recall(folderDir);
+    return next()
+  }
   },
 
-  postNestedFolder: function(req, res, next){
-    console.log('inside postFolder');
 
-    console.log
-
-  },
 
   deleteFolder: function (req, res, next) {
     const folderDir = 'server/ExportFolder/nextsketch';
@@ -89,9 +113,9 @@ const fileController = {
         const name = `${folderDir}/${file}`;
 
         //skip reading over all the files inside node_modules for efficiency
-        if (file === 'node_modules') {
+        if (name === 'node_modules') {
           //but if we click to delete it, delete it
-          fs.rmSync(name, { recursive: true });
+          if(req.body.name === 'node_modules') fs.rmSync(name, { recursive: true });
           continue;
         }
 
@@ -155,6 +179,27 @@ const fileController = {
 
       }
   });
+
+  // async function copyFilesAndDirectories(source, target) {
+  //   fs.ensureDirSync(target);
+  
+  //   const filesAndDirectories = glob.sync('**/*', { cwd: source, nodir: true });
+  
+  //   for (const item of filesAndDirectories) {
+  //     const sourcePath = path.join(source, item);
+  //     const targetPath = path.join(target, item);
+  
+  //     fs.ensureDirSync(path.dirname(targetPath));
+  //     fs.copyFileSync(sourcePath, targetPath);
+  //   }
+  // }
+  
+  // try {
+  //   copyFilesAndDirectories(sourceDir, path.join(targetDir, 'nextsketch'));
+  //   console.log('Directory and its contents copied successfully.');
+  // } catch (err) {
+  //   console.error(`Error copying directory: ${err}`);
+  // }
 
 
     return next();
